Move specific components to 'components.ex' in the dir that needs them
