+++ elixir/apps/web/lib/web/live/settings_live/identity_providers/new/oidc.ex
@@ -1,6 +1,6 @@
 defmodule Web.SettingsLive.IdentityProviders.New.OIDC do
   use Web, :live_view
-  import Web.FormComponents
+  import Web.SettingsLive.IdentityProviders.New.Components
 
   # TODO: Use a changeset for this
   @form_initializer %{
